feat: detailView 리뷰 구현

+++ src/main/java/poris/fruitlight/dto/ReviewInfo.java
@@ -8,7 +8,7 @@ import lombok.Data;
  */
 @Data
 public class ReviewInfo {
-	private int starRateAvg;				// 리뷰 평균 점수
+	private float starRateAvg;				// 리뷰 평균 점수
 	private float totalReviewScore;			// 리뷰 총 점수
 	private int reviewCount;
 }
